Separate AstTypes.h into multiple parts.
Just machanical changes.
