--- parser/AstTypes.cpp
@@ -1,12 +1,0 @@
-/* Copyright (c) 2018 - present, VE Software Inc. All rights reserved
- *
- * This source code is licensed under Apache 2.0 License
- *  (found in the LICENSE.Apache file in the root directory)
- */
-#include "parser/AstTypes.h"
-
-namespace vesoft {
-
-
-
-}
