+++ parser/TraverseSentences.cpp
@@ -1,0 +1,85 @@
+/* Copyright (c) 2018 - present, VE Software Inc. All rights reserved
+ *
+ * This source code is licensed under Apache 2.0 License
+ *  (found in the LICENSE.Apache file in the root directory)
+ */
+#include "parser/TraverseSentences.h"
+
+namespace vesoft {
+
+std::string GoSentence::toString() const {
+    std::string buf;
+    buf.reserve(256);
+    buf += "GO ";
+    if (stepClause_ != nullptr) {
+        buf += stepClause_->toString();
+    }
+    if (fromClause_ != nullptr) {
+        buf += " ";
+        buf += fromClause_->toString();
+    }
+    if (overClause_ != nullptr) {
+        buf += " ";
+        buf += overClause_->toString();
+    }
+    if (whereClause_ != nullptr) {
+        buf += " ";
+        buf += whereClause_->toString();
+    }
+    if (returnClause_ != nullptr) {
+        buf += " ";
+        buf += returnClause_->toString();
+    }
+
+    return buf;
+}
+
+std::string MatchSentence::toString() const {
+    return "MATCH sentence";
+}
+
+std::string UseSentence::toString() const {
+    return "USE NAMESPACE " + *ns_;
+}
+
+std::string SetSentence::toString() const {
+    std::string buf;
+    buf.reserve(256);
+    buf = "(";
+    buf += left_->toString();
+    switch (op_) {
+        case UNION:
+            buf += " UNION ";
+            break;
+        case INTERSECT:
+            buf += " INTERSECT ";
+            break;
+        case MINUS:
+            buf += " MINUS ";
+            break;
+    }
+    buf += right_->toString();
+    buf += ")";
+    return buf;
+}
+
+std::string PipedSentence::toString() const {
+    std::string buf;
+    buf.reserve(256);
+    buf += left_->toString();
+    buf += " | ";
+    buf += right_->toString();
+    return buf;
+}
+
+std::string AssignmentSentence::toString() const {
+    std::string buf;
+    buf.reserve(256);
+    buf += "$";
+    buf += *variable_;
+    buf += " = ";
+    buf += sentence_->toString();
+    return buf;
+}
+
+}   // namespace vesoft
