+++ parser/parser.yy
@@ -13,7 +13,7 @@
 #include <iostream>
 #include <sstream>
 #include <string>
-#include "AstTypes.h"
+#include "parser/Statement.h"
 
 namespace vesoft {
